Bug fix for adjustment of raw predictions
To mark the predictions after the first positive prediction for each machine ID as 1.0, the index of the first positive prediction and the last index of the data for the machine was being used. While setting the values to 1.0, I changed to slice of data accessed from [first_pos_idx: last_pos_idx+1] to [first_pos_idx:last_pos_idx] as the last index is included in the data returned (unlike in python list slicing).